consolidating shear with no rotation tests; simplifying tilt based tests
